Slice chapters into sections, show individually.
This brings over the functionality in $scope.chapterLoaded.

This includes the quirk that the first section is always shown with
preceding material (usually a short introductory blurb - not to be
confused with a separate introduction document).

Still need to prevent redundant fetching by stripping anchor tags, and
perhaps rethink how the cache works. (Probably better to cache by full
URL, including anchor, or by 'location', rather than the current nested
structure of URL -> array of sections.)