Improve TOC styling, remove it from content area.
Table of Contents text color now matches the rest of the design (instead
of being a separate link color); also made 2nd popover theme-compliant.

The initial page is now the welcome page. This change takes us from 3
ways to access the Table of Contents down to 2, and eliminates several
special cases that were in place for treating the TOC like other pages.

Also, manually 'namespaced' the stuff in the curriculum theme CSS so it
only affects things under #curriculum.