Fix intra-curricular links, refactor processing.
Previously, cross-references were implemented by adding onclick strings
to the HTML, to invoke angular functions. This took place during
curriculum preprocessing in Python.

Now, this happens in the client with the other DOM processing steps, and
the event handlers are added by JS (and thus not visible in the HTML).
This required changing how we process and load the curriculum pages.
Also, this change finally moves the processing out of
fetchContent.fulfilled, as I promised @ttsuchiya a while ago, and it
removes the now-defunct maximization code from curriculumState.