Filter questions by language immediately.
Fixes a bug in which both versions of a question (Python & Javascript)
would show up the first time the questions page was viewed. This occured
because the relevant elements had not yet been generated at the time of
viewing.

In the future, we might want to generate the question elements right
after fetching (at the same time as the other processing steps), which
would also allow us to omit the JS from the AsciiDoc files and write the
questions in a more plain format.