Fix curriculum search crash.
Our hilitor.js was using an array in place of an object.
This is not only inappropriate, but also dangerous:
if it tried to highlight "length", it would assign to Array.length,
which in JS has the effect of resizing the array!
Hence, the bug manifested specifically upon searching for "length".
(But it also could also be problematic if this tried to highlighting a
number; assigning to arr["1000000"] has the effect of extending the
array to include the index 1000000 in JS!)