kludge fix for pitchshift
Fixes setEffect(2, PITCHSHIFT, PITCHSHIFT_SHIFT, -4.0)

Fixes GTCMT/earsketch#2679

The new fixEffects(result) method in runner.ts expects result.length to be the song length in measures, but instead it is set to the default value of 0. This is because recent changes for setTempo() caused a re-ordering of the runner.ts postRun() logic. Because of this there is a missing "end" point in the pitchshift env in pitchshiftClips().