expect err msg to be an array of strings
This may be related to the new "X" and "!" icons for error and warning messages. The broken code was expecting a string, but `msg.text` is actually an array of strings, for py and js, such as `[0: "SyntaxError: There is an error with the syntax (or arrangement) of code", 1: " Unexpected token (4", 2: "0) on line 4"]`.